+++ packages/mds-geography-author/tests/api.spec.ts
@@ -513,7 +513,7 @@ describe('Tests app', () => {
         .expect(400)
         .end((err, result) => {
           test.assert(result.body.error.name === `ValidationError`)
-          test.assert(result.body.error.reason.includes('publish_date'))
+          test.assert(result.body.error.info[0].params.additionalProperty.includes('publish_date'))
           test.value(result).hasHeader('content-type', APP_JSON)
           done(err)
         })
@@ -533,7 +533,7 @@ describe('Tests app', () => {
         .expect(400)
         .end((err, result) => {
           test.assert(result.body.error.name === `ValidationError`)
-          test.assert(result.body.error.reason.includes('publish_date'))
+          test.assert(result.body.error.info[0].params.additionalProperty.includes('publish_date'))
           test.value(result).hasHeader('content-type', APP_JSON)
           done(err)
         })