+++ packages/mds-ingest-service/repository/index.ts
@@ -34,6 +34,8 @@ import {
   EventDomainModel,
   GetDevicesOptions,
   GetDevicesResponse,
+  GetEventsWithDeviceAndTelemetryInfoOptions,
+  GetEventsWithDeviceAndTelemetryInfoResponse,
   GetVehicleEventsFilterParams,
   GetVehicleEventsOrderOption,
   GetVehicleEventsResponse,
@@ -49,32 +51,46 @@ import { TelemetryEntity, TelemetryEntityModel } from './entities/telemetry-enti
 import {
   DeviceDomainToEntityCreate,
   DeviceEntityToDomain,
+  DeviceEntityToDomainWithIdentityColumn,
   EventAnnotationDomainToEntityCreate,
   EventAnnotationEntityToDomain,
   EventDomainToEntityCreate,
   EventEntityToDomain,
-  EventEntityToDomainWithIdentity,
+  EventEntityToDomainWithIdentityColumn,
+  EventWithDeviceAndTelemetryInfoEntityToDomain,
   MigratedDeviceToEntityCreate,
   MigratedEventToEntityCreate,
   MigratedTelemetryToEntityCreate,
   TelemetryDomainToEntityCreate,
-  TelemetryEntityToDomain
+  TelemetryEntityToDomain,
+  TelemetryEntityToDomainWithIdentityColumn
 } from './mappers'
 import migrations from './migrations'
 import { MigratedEntityModel } from './mixins/migrated-entity'
+import views from './views'
+import { EventWithDeviceAndTelemetryInfoEntity } from './views/event-with-device-and-telemetry-info'
 
-type GetDeviceQueryParams = GetDevicesOptions & Cursor
-type VehicleEventsQueryParams = GetVehicleEventsFilterParams & Cursor
+type WithCursorOptions<P extends object> = P & Cursor
+
+const buildCursor = <T extends {}>(options: WithCursorOptions<T>) =>
+  Buffer.from(JSON.stringify(options), 'utf-8').toString('base64')
+
+const parseCursor = <T extends {}>(cursor: string): WithCursorOptions<T> => {
+  try {
+    return JSON.parse(Buffer.from(cursor, 'base64').toString('utf-8'))
+  } catch (error) {
+    throw new ValidationError('Invalid cursor', error)
+  }
+}
 
 class IngestReadWriteRepository extends ReadWriteRepository {
   constructor() {
-    super('ingest', { entities, migrations })
+    super('ingest', { entities: [...entities, ...views], migrations })
   }
 
   public createEvents = async (events: EventDomainCreateModel[]) => {
-    const { connect } = this
     try {
-      const connection = await connect('rw')
+      const connection = await this.connect('rw')
       const { raw: entities }: InsertReturning<EventEntity> = await connection
         .getRepository(EventEntity)
         .createQueryBuilder()
@@ -89,9 +105,8 @@ class IngestReadWriteRepository extends ReadWriteRepository {
   }
 
   public createTelemetries = async (events: TelemetryDomainCreateModel[]) => {
-    const { connect } = this
     try {
-      const connection = await connect('rw')
+      const connection = await this.connect('rw')
       const { raw: entities }: InsertReturning<TelemetryEntity> = await connection
         .getRepository(TelemetryEntity)
         .createQueryBuilder()
@@ -106,9 +121,8 @@ class IngestReadWriteRepository extends ReadWriteRepository {
   }
 
   public createDevices = async (events: DeviceDomainCreateModel[]) => {
-    const { connect } = this
     try {
-      const connection = await connect('rw')
+      const connection = await this.connect('rw')
       const { raw: entities }: InsertReturning<DeviceEntity> = await connection
         .getRepository(DeviceEntity)
         .createQueryBuilder()
@@ -125,9 +139,8 @@ class IngestReadWriteRepository extends ReadWriteRepository {
   public createEventAnnotations = async (
     eventAnnotations: EventAnnotationDomainCreateModel[]
   ): Promise<EventAnnotationDomainModel[]> => {
-    const { connect } = this
     try {
-      const connection = await connect('rw')
+      const connection = await this.connect('rw')
       const { raw: entities }: InsertReturning<EventAnnotationEntity> = await connection
         .getRepository(EventAnnotationEntity)
         .createQueryBuilder()
@@ -153,22 +166,11 @@ class IngestReadWriteRepository extends ReadWriteRepository {
     }
   }
 
-  private buildGetDevicesCursor = (params: GetDeviceQueryParams) =>
-    Buffer.from(JSON.stringify(params), 'utf-8').toString('base64')
-
-  private parseGetDevicesCursor = (cursor: string): GetDeviceQueryParams => {
-    try {
-      return JSON.parse(Buffer.from(cursor, 'base64').toString('utf-8'))
-    } catch (error) {
-      throw new ValidationError('Invalid cursor', error)
-    }
-  }
-
   private getDevicesQuery = async ({
     limit = 100,
     beforeCursor,
     afterCursor
-  }: GetDeviceQueryParams): Promise<GetDevicesResponse> => {
+  }: WithCursorOptions<GetDevicesOptions>): Promise<GetDevicesResponse> => {
     try {
       const connection = await this.connect('ro')
       const query = connection.createQueryBuilder(DeviceEntity, 'device')
@@ -195,10 +197,10 @@ class IngestReadWriteRepository extends ReadWriteRepository {
         cursor: {
           next:
             nextAfterCursor &&
-            this.buildGetDevicesCursor({ ...cursor, beforeCursor: null, afterCursor: nextAfterCursor }),
+            buildCursor<GetDevicesOptions>({ ...cursor, beforeCursor: null, afterCursor: nextAfterCursor }),
           prev:
             nextBeforeCursor &&
-            this.buildGetDevicesCursor({ ...cursor, beforeCursor: nextBeforeCursor, afterCursor: null })
+            buildCursor<GetDevicesOptions>({ ...cursor, beforeCursor: nextBeforeCursor, afterCursor: null })
         }
       }
     } catch (error) {
@@ -210,10 +212,11 @@ class IngestReadWriteRepository extends ReadWriteRepository {
     this.getDevicesQuery({ ...options, beforeCursor: null, afterCursor: null })
 
   public getDevicesUsingCursor = async (cursor: string): Promise<GetDevicesResponse> =>
-    this.getDevicesQuery(this.parseGetDevicesCursor(cursor))
+    this.getDevicesQuery(parseCursor<GetDevicesOptions>(cursor))
 
-  private getEvents = async (params: VehicleEventsQueryParams): Promise<GetVehicleEventsResponse> => {
-    const { connect } = this
+  private getEvents = async (
+    params: WithCursorOptions<GetVehicleEventsFilterParams>
+  ): Promise<GetVehicleEventsResponse> => {
     const {
       time_range,
       geography_ids,
@@ -232,7 +235,7 @@ class IngestReadWriteRepository extends ReadWriteRepository {
     } = params
 
     try {
-      const connection = await connect('ro')
+      const connection = await this.connect('ro')
 
       const query = connection
         .getRepository(EventEntity)
@@ -336,23 +339,24 @@ class IngestReadWriteRepository extends ReadWriteRepository {
       return {
         events: data.map(EventEntityToDomain.mapper()),
         cursor: {
-          next: nextAfterCursor && this.buildCursor({ ...cursor, beforeCursor: null, afterCursor: nextAfterCursor }),
-          prev: nextBeforeCursor && this.buildCursor({ ...cursor, beforeCursor: nextBeforeCursor, afterCursor: null })
+          next:
+            nextAfterCursor &&
+            buildCursor<GetVehicleEventsFilterParams>({
+              ...cursor,
+              beforeCursor: null,
+              afterCursor: nextAfterCursor
+            }),
+          prev:
+            nextBeforeCursor &&
+            buildCursor<GetVehicleEventsFilterParams>({
+              ...cursor,
+              beforeCursor: nextBeforeCursor,
+              afterCursor: null
+            })
         }
       }
     } catch (error) {
       throw RepositoryError(error)
-    }
-  }
-
-  private buildCursor = (cursor: VehicleEventsQueryParams) =>
-    Buffer.from(JSON.stringify(cursor), 'utf-8').toString('base64')
-
-  private parseCursor = (cursor: string): VehicleEventsQueryParams & { limit: number } => {
-    try {
-      return JSON.parse(Buffer.from(cursor, 'base64').toString('utf-8'))
-    } catch (error) {
-      throw new ValidationError('Invalid cursor', error)
     }
   }
 
@@ -423,7 +427,7 @@ class IngestReadWriteRepository extends ReadWriteRepository {
     this.getEvents({ ...options, beforeCursor: null, afterCursor: null, limit: options.limit ?? 100 })
 
   public getEventsUsingCursor = async (cursor: string): Promise<GetVehicleEventsResponse> =>
-    this.getEvents(this.parseCursor(cursor))
+    this.getEvents(parseCursor<GetVehicleEventsFilterParams>(cursor))
 
   public getTripEvents = async (params: ReadTripEventsQueryParams) => {
     const { skip, take = 100, start_time, end_time, provider_id } = params
@@ -473,18 +477,22 @@ class IngestReadWriteRepository extends ReadWriteRepository {
   public getLatestTelemetryForDevices = async (device_ids: UUID[]): Promise<TelemetryDomainModel[]> => {
     try {
       const connection = await this.connect('ro')
-      const entities = await connection
-        .createQueryBuilder(TelemetryEntity, 'telemetry')
-        .innerJoin(
-          subquery =>
-            subquery
-              .select('id, RANK() OVER (PARTITION BY device_id ORDER BY timestamp DESC) AS rownum')
-              .where('device_id = ANY(:device_ids)', { device_ids })
-              .from(TelemetryEntity, 't'),
-          'last_device_telemetry',
-          'last_device_telemetry.id = telemetry.id AND last_device_telemetry.rownum = 1'
-        )
-        .getMany()
+
+      /**
+       * Run the query within a transaction so we can safely disable bitmapscan
+       * https://github.com/typeorm/typeorm/blob/master/docs/transactions.md
+       */
+      const entities = await connection.transaction(async manager => {
+        await manager.query(`select set_config('enable_bitmapscan','off', true)`)
+        return await manager
+          .createQueryBuilder(TelemetryEntity, 'telemetry')
+          .distinctOn(['device_id'])
+          .orderBy('device_id', 'DESC')
+          .addOrderBy('timestamp', 'DESC')
+          .where('device_id = ANY (:device_ids)', { device_ids })
+          .getMany()
+      })
+
       return entities.map(TelemetryEntityToDomain.mapper())
     } catch (error) {
       throw RepositoryError(error)
@@ -494,7 +502,7 @@ class IngestReadWriteRepository extends ReadWriteRepository {
   public writeMigratedDevice = async (
     devices: Array<Device & Required<RecordedColumn>>,
     migrated_from: MigratedEntityModel
-  ): Promise<DeviceDomainModel[]> => {
+  ): Promise<Array<DeviceDomainModel & IdentityColumn>> => {
     try {
       const connection = await this.connect('rw')
       const { raw: entities }: InsertReturning<DeviceEntityModel> = await connection
@@ -508,7 +516,7 @@ class IngestReadWriteRepository extends ReadWriteRepository {
         )
         .returning('*')
         .execute()
-      return entities.map(DeviceEntityToDomain.mapper())
+      return entities.map(DeviceEntityToDomainWithIdentityColumn.mapper())
     } catch (error) {
       throw RepositoryError(error)
     }
@@ -528,7 +536,7 @@ class IngestReadWriteRepository extends ReadWriteRepository {
         .onConflict('DO NOTHING')
         .returning('*')
         .execute()
-      return entities.map(EventEntityToDomainWithIdentity.mapper())
+      return entities.map(EventEntityToDomainWithIdentityColumn.mapper())
     } catch (error) {
       throw RepositoryError(error)
     }
@@ -537,7 +545,7 @@ class IngestReadWriteRepository extends ReadWriteRepository {
   public writeMigratedTelemetry = async (
     telemetry: Array<Telemetry & Required<RecordedColumn>>,
     migrated_from: MigratedEntityModel
-  ): Promise<TelemetryDomainModel[]> => {
+  ): Promise<Array<TelemetryDomainModel & IdentityColumn>> => {
     try {
       const connection = await this.connect('rw')
       const { raw: entities }: InsertReturning<TelemetryEntityModel> = await connection
@@ -548,11 +556,82 @@ class IngestReadWriteRepository extends ReadWriteRepository {
         .onConflict('DO NOTHING')
         .returning('*')
         .execute()
-      return entities.map(TelemetryEntityToDomain.mapper())
+      return entities.map(TelemetryEntityToDomainWithIdentityColumn.mapper())
     } catch (error) {
       throw RepositoryError(error)
     }
   }
+
+  private getEventsWithDeviceAndTelemetryInfo = async ({
+    provider_ids,
+    device_ids,
+    time_range,
+    limit = 100,
+    ...cursor
+  }: WithCursorOptions<GetEventsWithDeviceAndTelemetryInfoOptions>): Promise<GetEventsWithDeviceAndTelemetryInfoResponse> => {
+    try {
+      const connection = await this.connect('ro')
+
+      const query = connection.getRepository(EventWithDeviceAndTelemetryInfoEntity).createQueryBuilder('events')
+
+      if (provider_ids?.length) {
+        query.andWhere('provider_id = ANY(:provider_ids)', { provider_ids })
+      }
+
+      if (device_ids?.length) {
+        query.andWhere('device_id = ANY(:device_ids)', { device_ids })
+      }
+
+      if (time_range?.start) {
+        query.andWhere('timestamp >= :start', time_range)
+      }
+
+      if (time_range?.end) {
+        query.andWhere('timestamp <= :end', time_range)
+      }
+
+      const {
+        data: entities,
+        cursor: { beforeCursor: nextBeforeCursor, afterCursor: nextAfterCursor }
+      } = await buildPaginator({
+        alias: query.alias,
+        entity: EventWithDeviceAndTelemetryInfoEntity,
+        query: {
+          limit,
+          order: 'ASC',
+          afterCursor: cursor.afterCursor ?? undefined,
+          beforeCursor: cursor.beforeCursor ?? undefined
+        },
+        paginationKeys: ['id']
+      }).paginate(query)
+
+      const options = { limit, time_range, provider_ids, device_ids }
+
+      return {
+        events: entities.map(EventWithDeviceAndTelemetryInfoEntityToDomain.mapper()),
+        cursor: {
+          next:
+            nextAfterCursor &&
+            buildCursor<GetDevicesOptions>({ ...options, beforeCursor: null, afterCursor: nextAfterCursor }),
+          prev:
+            nextBeforeCursor &&
+            buildCursor<GetDevicesOptions>({ ...options, beforeCursor: nextBeforeCursor, afterCursor: null })
+        }
+      }
+    } catch (error) {
+      throw RepositoryError(error)
+    }
+  }
+
+  public getEventsWithDeviceAndTelemetryInfoUsingOptions = async (
+    options: GetEventsWithDeviceAndTelemetryInfoOptions
+  ): Promise<GetEventsWithDeviceAndTelemetryInfoResponse> =>
+    this.getEventsWithDeviceAndTelemetryInfo({ ...options, beforeCursor: null, afterCursor: null })
+
+  public getEventsWithDeviceAndTelemetryInfoUsingCursor = async (
+    cursor: string
+  ): Promise<GetEventsWithDeviceAndTelemetryInfoResponse> =>
+    this.getEventsWithDeviceAndTelemetryInfo(parseCursor<GetEventsWithDeviceAndTelemetryInfoOptions>(cursor))
 
   /**
    * Nukes everything from orbit. Boom.