+++ packages/mds-transaction-service/repository/index.ts
@@ -15,21 +15,17 @@
  */
 
 import { InsertReturning, ReadWriteRepository, RepositoryError } from '@mds-core/mds-repository'
-import { schemaValidator } from '@mds-core/mds-schema-validators'
 import { UUID } from '@mds-core/mds-types'
 import { NotFoundError } from '@mds-core/mds-utils'
-import Joi from 'joi'
 import { Between, Brackets, FindOperator, In, LessThan, MoreThan } from 'typeorm'
 import { buildPaginator, Cursor } from 'typeorm-cursor-pagination'
 import {
-  FEE_TYPE,
-  SORTABLE_COLUMN,
-  SORT_DIRECTION,
   TransactionDomainModel,
   TransactionOperationDomainModel,
   TransactionSearchParams,
   TransactionStatusDomainModel
 } from '../@types'
+import { validateTransactionSearchParams } from '../service/validators'
 import { TransactionOperationEntity } from './entities/operation-entity'
 import { TransactionStatusEntity } from './entities/status-entity'
 import { TransactionEntity } from './entities/transaction-entity'
@@ -51,27 +47,6 @@ const testEnvSafeguard = () => {
     throw new Error(`This method is only supported when executing tests`)
   }
 }
-
-const { validate: validateTransactionSearchParams } = schemaValidator<TransactionSearchParams>(
-  Joi.object<TransactionSearchParams>()
-    .keys({
-      provider_id: Joi.string().uuid(),
-      start_timestamp: Joi.number().integer(),
-      end_timestamp: Joi.number().integer(),
-      search_text: Joi.string(),
-      start_amount: Joi.number(),
-      end_amount: Joi.number(),
-      fee_type: Joi.string().allow(...FEE_TYPE),
-      before: Joi.string(),
-      after: Joi.string(),
-      limit: Joi.number().integer().min(1).max(1000).default(10),
-      order: Joi.object<TransactionSearchParams['order']>().keys({
-        column: Joi.string().allow(...SORTABLE_COLUMN),
-        direction: Joi.string().allow(...SORT_DIRECTION)
-      })
-    })
-    .unknown(false)
-)
 
 class TransactionReadWriteRepository extends ReadWriteRepository {
   public getTransaction = async (transaction_id: UUID): Promise<TransactionDomainModel> => {